Updated Readme
I discovered the error with updating the databases arised from having multiple
Entiry Frameworks installed. Specifying which one to use fixes the database update command.
I updated the readme to reflect this and changed the name of the Admin Account Context
to make it easier to run the commands

signed-off-by John Brusaw <jbrusaw@pdx.edu>