Refactored form rules to enable more dynamic rule generation.
Added validation of the service delivery table: invalid entries appear in
red, bolded text with a red background and block form submission.

Added functions for comparing and parsing dates and times.

Added comparison between provider&employer sign dates and form date
fields.

Appended form errors to the bottom of the form after trying to submit.

Added functionality for packaging form field data to a format that can
be submitted to the backend.

Signed-off-by: Lenovo Windows - Yves Wienecke <whyves@icloud.com>