Added Roles to AdminUI
I replaced the existing Admin Account scaffolded code with a new one
that gives us the ability to add custom fields to accounts. Now when
users register, than can enter their name and choose a role: Employee or
Administrator. I also moved a couple files around for refactoring
purporses, and changed the Readme to reflect the new databse changes.

You will need to delete your existing AzureDB and run the
Update-Database commands in the Readme for this to work

Note that to test it, you'll have to create an account and then either
go to the account page, download your data, and see your name/role, or use
the SQL Server Object explorer to see names/roles in the localdb

Signed-off-by: John Brusaw <jwbrusaw@gmail.com>