Redid Roles and Added Lock Table View
I realized the way I did roles earlier was sloppy and ineffective, so I
redid the whole thing to make better use of ASP.NET built-in roles. Now
when someone isn't logged in, they're redirected to the login page. In
addition, there is now a Lock Table which only administrators can
access. I also made a few adjustments to the registration and login
process to make it easier, and I also removed all the password
requirements. This will be put back in when we release, but I was
getting real tired of creating fake accounts with complex passwords
The database is also seeded with two accounts to test on. See Readme
for usernames and passwords

Signed-off-by: John Brusaw <jwbrusaw@gmail.com>