Added Mass Download PDF function
Added a button on the main page to download every PDF matching the
selected criteria

As of now, it's very inefficent. It takes ~2 seconds to generate a PDF,
of which the vast majority of that time is spent adding the images of
the timesheets. For 1,600 sheets, that means it'll take ~45 minutes to
render and download them all. I spent about an hour researching
improvements, but all I could come up with is spending money to get a
better PDF Renderer : /

I highly recommend filtering your sheets to 2-5 before attempting to
download. Otherwise you'll be waiting awhile

Signed-off-by: John Brusaw <jwbrusaw@gmail.com>