Added Pay Periods
Added Pay Periods and gave Administrators the ability to create,
update, delete, and start pay periods. There is a global
"CurrentPayPeriod" variable that determines the default to/from dates
when filtering sheets, as well as a warning that displays to all users
when running low on dates.

This required another Database Context, so please run "Update-Database
-Context PayPeriodContext" to enable functionality

Signed-off-by: John Brusaw <jwbrusaw@gmail.com>