Reorganized AbstractFormObject and Derived Classes
I moved common fields into the base class. This required an update
to the expected json due to how Newtonsoft defaults to derived class
first, then base class, for serializing properties.

Along with this the FromTextract method has been overhauled to take
the formtype then works with a protected version to create the
correct form and return it.

Signed-off-by: David Post <post@pdx.edu>