Added guard against false positive on front page
To allow more edits we need to loosen the tolerance, but to protect
against false positive we want to double check that this isn't the back
page. So we verify that it has something that should be only on the
front page, and then check that it doesn't contain something only on the
back page.

Signed-off-by: David Post <post@pdx.edu>