Improved Query Functionality
Previously, each function got all the timesheets from the database and
filtered the list down 1 filter at a time.

Now, a single function builds the query up and queries the database
once, which should be significantly faster. This is also better for
scalability and code maintenance down the road.

I also renamed a couple variables in HomeController to make Curtis
happy. There are a lot more variables that need improvement, but that's
another ticket for another time

Signed-off-by: John Brusaw <jwbrusaw@gmail.com>