Synchronizing the staging and submission ids
It's a little more complicated than I would like, but it's workable in
normal situations. When running through our test endpoints we do run
into the issue of overlapping submissions, but in practice that should
not be the case, and we can also change the way we handle this in
testing to accomadate the normal workflow by having SubmitTest take
the staging we want to submit and copy it in the staging table.

Signed-off-by: David Post <post@pdx.edu>